+++ pybamm/models/submodels/interface/kinetics/butler_volmer.py
@@ -36,7 +36,7 @@ class ButlerVolmer(BaseKinetics):
         return 2 * j0 * pybamm.sinh(prefactor * eta_r)
 
     def _get_dj_dc(self, variables):
-        """ See :meth:`pybamm.interface.kinetics.BaseKinetics._get_dj_dc` """
+        """See :meth:`pybamm.interface.kinetics.BaseKinetics._get_dj_dc`"""
         c_e, delta_phi, j0, ne, ocp, T = self._get_interface_variables_for_first_order(
             variables
         )
@@ -47,7 +47,7 @@ class ButlerVolmer(BaseKinetics):
         )
 
     def _get_dj_ddeltaphi(self, variables):
-        """ See :meth:`pybamm.interface.kinetics.BaseKinetics._get_dj_ddeltaphi` """
+        """See :meth:`pybamm.interface.kinetics.BaseKinetics._get_dj_ddeltaphi`"""
         _, delta_phi, j0, ne, ocp, T = self._get_interface_variables_for_first_order(
             variables
         )